move metadata stuff to metadata folder
