+++ csharp/destinations/snowflake/Quix.Snowflake.Infrastructure/Metadata/EventGroupRepository.cs
@@ -1,24 +1,24 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Quix.Snowflake.Domain.Models;
 using Quix.Snowflake.Domain.Repositories;
 
-namespace Quix.Snowflake.Infrastructure.TimeSeries.Repositories
+namespace Quix.Snowflake.Infrastructure.Metadata
 {
-    public class ParameterRepository : IParameterRepository
+    public class EventGroupRepository : IEventGroupRepository
     {
-        public Task Save(ICollection<TelemetryParameter> parameters)
+        public Task Save(ICollection<TelemetryEventGroup> eventGroups)
         {
             throw new System.NotImplementedException();
         }
 
-        public IQueryable<TelemetryParameter> GetAll()
+        public IQueryable<TelemetryEventGroup> GetAll()
         {
             throw new System.NotImplementedException();
         }
 
-        public Task<long> Delete(string streamId, ICollection<string> parameterIds)
+        public Task<long> Delete(string streamId, ICollection<string> groupPaths)
         {
             throw new System.NotImplementedException();
         }
@@ -28,7 +28,7 @@ namespace Quix.Snowflake.Infrastructure.TimeSeries.Repositories
             throw new System.NotImplementedException();
         }
 
-        public Task BulkWrite(IEnumerable<TelemetryParameter> requests)
+        public Task BulkWrite(IEnumerable<TelemetryEventGroup> requests)
         {
             throw new System.NotImplementedException();
         }
